module.c: Initialize variables before use
Yeah, the way the code is currently structured, the code won't
(shouldn't) use uninitialized variables, but gcc doesn't know that.

In general, leaving pointers pointing to uninitialized memory and just
hoping that the user won't touch them is a terrible practice, so I'm not
too upset with gcc for complaining.

Change-Id: I03f268f06313fc71a2b2cde6c579b19b7be17d7a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>