Fix problem where oconf binary was installed as "oconf"
This fixes a regression introduced by the autotools machinery, due to
me accidentally installing the binary that was supposed to be known as
"-oconf" as "oconf".

Because, seriously, you do NOT use "-" as a prefix for "hidden"
binaries! There are not many conventions universally followed by unix
tools. That any argument to a tool that begins with a "-" is a binary is
one of the few - so naming files that way makes globbing in that
directory potentially lethal, and you may not disagree with this until
you've executed the following in your shell:

 touch /-rf /--no-preserve-root
 cd /
 rm -i * # it's safe, because there's a -i, amiright?

And while I'm ranting, we already have a directory full of binaries for
mon to use - just put the bloody thing there, and mon will magically not
consider it a subcommand.

Change-Id: I52bb58cac9d6ddc3ac85025bf3ab303e8e800b8c
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>