shared: Fix use_syslog definition
I actually developed this to use log_to_syslog as a variable name. I
then decided that it seemed to make sense to keep the variable name
consistent with nagios to make it more memorable, so I proceeded to
rename it. Without testing properly, apparently - log variables that
aren't prefixed with "log_" don't work.

Add support for the nagios variable name, for real.

Change-Id: Iefca368ea21f659678eb32aff6e1fa26728c9538
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>