install: Make merlin run as non-root
This fixes bug #4481.

It's a bit messy in places - because the mon suite is very, very
hardcoded in places, a lot of scripts end up with a ".in" suffix so we
can generate them with (hopefully) useful settings. A general
frameworkification of the mon suite would be helpful to avoid these
things.

In addition, the makefile is looking slightly horrible, as wildcards
don't seem to match generated sources, unless they've already been
generated, and then the wildcard starts to match them an extra time,
so either way, either rebuilds or fresh builds fail with the
wildcards, as far as I can tell. Sigh, I'm sorry.

Change-Id: Ic34d1b53c1e88e8bed4ad257de087acf7203cb38
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>