net: Fix handling of semi-connected nodes
We open connections non-blockingly, which is nice, but then we
completely forget to care about waiting for them to become connected,
which is not nice, and we end up with weird-ass debug logs like

[1412341434] 6: NODESTATE: 172.27.102.204: STATE_NONE -> STATE_PENDING: Connecting
[1412341434] 3: Failed to read from socket 9 into 0x225b0a0 for peer node 172.27.102.204: No route to host
[1412341434] 7: bytes_read: -1; errno: 113; strerror(113): No route to host

read() very rarely return "No route to host", so something is
definitely fishy. The trick is to pay attention to EINPROGRESS in
connect()'s manpage: we need to wait for writability before doing
anything with the socket, but we only waited for readability for any
non-binlogged socket.

Change-Id: Ic65e15c72a983629ce0a0836749f7d92552f8795
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>