net: Fix handling of incoming connections during negotiation
The code claimed that it was most likely a bug if two connections were
established at once. However, it has proved possible (bug #9291)
produce this case.

When that happened, we added a special case not to re-send a
CTRL_ACTIVE to it. In all other cases, we tend to go with the "an
extra CTRL_ACTIVE never hurt anyone", but here we decided that we
should bail - and it seems it actually matters.

When that happened, we didn't end up getting a CTRL_ACTIVE on the remote
module side at all. Merlin's handshake is not a proper, three-way one
- we only wait for a CTRL_ACTIVE from the remote side before starting
to send data.

This situation resulted in two nodes, where one of the modules
considers the other to be connected, while the second consider the
first one to be disconnected.

This should fix bug #9291.

Change-Id: I131b2ad823ce85dabbd15e1c8a224d857001d1bd
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>