Remove python implementation of object configuration splitter
There are bugs in it that doesn't exist in the C version provided
by Naemon, and it's far less efficient than the C version, so
we might as well get rid of it.

While at it, we change the path to the oconfsplit binary to be
set by the user. Since people can override it with a commandline
switch anyway, nonstandard configurations should also work
well, while making the default ones work out of the box.

Change-Id: Idda67f8e7c929e6f70982dd15b9b74c62fb22237
Signed-off-by: Andreas Ericsson <ae@op5.se>