daemon: Make logging of executed commands debug-level
Since they sometimes contain sensitive information (like
passwords and stuff), logging them on the default loglevel
is a pretty bad idea. Users who do turn on debug-logging
really shouldn't be surprised when the logs contain a lot
of information that may be sensitive though, as that
informtation may be vital to tracking down potential
errors.

This fixes issue #9264.

Change-Id: I028e10b6f1c1e070c5f43626d3e04d8f3b901bdb
Signed-off-by: Andreas Ericsson <ae@op5.se>