mon oconf: Give 'push' a chance to find the oconfsplit binary
When we decided to put 'oconfsplit' in the user's $PATH and use
that to locate it, we never removed the check to see if it's
available, which meant config pushing only ever worked from
the same directory where oconfsplit is located (/usr/bin,
normally). That's pretty weird, so let's remove the check and
instead catch exceptions from the subprocess.Popen() call,
which will also catch other errors from the execution.
This ensures that "mon oconf push" works equally well no
matter which working directory the user happens to be using.

Change-Id: I4cb7c30183d85a97bcd2e8d6b9a889338a2a6d91
Signed-off-by: Andreas Ericsson <ae@op5.se>