daemon: Try to fix confused pollers more often
It's possible for pollers to believe they have newer config than the
master. However, that's not OK, and the master will ignore the poller
because of this, but the poller will not be receiving new config as
long as the master config is reported as being older, and the poller
can't push to the master, so they just agree to disagree and refuse to
talk again. That's not good.

Fix it by ignoring the mtime in the case where relations are
hierarchical - if the poller's config is different from what the
master expects, the master should push or the poller should poll,
always.

There was already a block against pollers pushing to masters, so this
line of thought had already crossed someones mind - this patch takes
the solution all the way.

This means that problems due to bug #9311 takes care of themselves,
eventually. However, as long as this decision is based on mtime rather
than hash, it will sometimes start out broken before fixing itself -
alas, the daemon does not know about pollers' expected hash, so this
fix is a lot simpler.

Change-Id: I4fab2c4a5a3e33231d9b8a57223ba0ca9603c2ab
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>