oconf: Nodesplit in module at startup
There are two things that must happen for a nodesplit to be requires:
 a) naemon must be restarted
 b) with new config

We've always focused on b), while focusing on a) makes things a lot
less complicated. And I care about Avril, and so should you.

That is, let the nodesplit run at naemon startup, with naemon's
already-parsed config - as the parsing happens once, instead of
once-per-node (with the old nodesplit function) or
once-per-node-squared (more or less, with the new nodesplit helper),
this could even end up being quicker: for a (actually proposed)
customer system with 1300 pollers, we would have been parsing the
config 1.6 million times (technically, we would have parsed the
object.cache, but it requires parsing nonetheless - and we do
postprocessing like verify the absence of cycles in it, anyway). Now
we do it once. Algoritms, yo.

It is not about performance, though: config parsing today is so quick,
you could even implement an obviously O(1) problem in O(n^2) and get
away with it, and config dumping happens every few seconds in every
naemon/nagios system that ever lived. If it was about performance, I
might try writing each poller group's file once (instead of once for
each poller - the old code actually did handle this case, so that's a
regression), hashing to detect if the config actually did change, or a
number of other obvious improvements.

Nay, this is about races: if we only write from one process - or, more
specifically, only write to one single file from one process - we
won't enter a race with the other processes that are busy writing the
same thing. We won't get corrupted config, meaning this fixes bugs
like #9311 (the cause, this time, not just the symptoms).

This thus imports the oconfsplit binary from naemon, kicks out all of
its command line parsing, uses the remains at startup, and removes all
the python nodesplit nonsense.

Change-Id: I4e1b3663c447ca22fe1e0dfc446759a6a8c6c6dc
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>