module: Reconnect when module goes away, not when binlog is full
node_send_event does not return whether the event was sent - it
returns whether the event was "handled" by the "node layer", i.e. if
it was either sent, or queued up by the binlogging functionality and
thus about to be sent later. If the daemon is backed up, it will still
be after a reconnect. If the daemon restarted, we should have
reconnected when it did that, not when our buffer is full.

So do precisely that - when the daemon disconnects from the module, we
close the old socket, and mark the module as disconnected, so
ipc_is_connected() will notice and reconnect for us.

Change-Id: Ib61286f943f1be518fadef8fa8e2392f2c2576d6
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>