Remove a bit of dead code
status.[ch] was never used. Ever. Like, "never" in terms of
"at no place in the codebase", but also, "never" in terms of "at least
not since 2009, and merlin was first released in 2010, so at least not
since the start of the common era". So we can just delete them, remove
them from the Makefile, and be done. Right?

I had you there, didn't I? No, you see, status.h is the only file
which references state.h, and we *do* use state.[ch]. So when you
merely remove status.h, in the belief that doing so will only remove
status.h, you get a compilation error, and given the awful record we
have at namespacing so far, fooling yourself that init_state is a
constructor from status.c (as opposed to being one from state.c) is
easy. So, when moving things around, I was fooled.

But now, I've actually figured it out, and this time, it won't get
away, meddling kids be damned!

Change-Id: I24557f419bfcd905e52f464874f2ccfeb019ef15
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>