Move single-use code away from "shared"
linked_item is defined in node.h, so it should not have it's basic API
defined anywhere but in the same file.

As this API is only used by node.c, I'm fine with it being defined
here - so move the add call into node.c as well, so it doesn't pollute
shared.

Change-Id: I53bc6ae0ee2de6040d71b3902c2594d87435935e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>