shared: Move configuration code to configuration file
shared.c is mostly composed of really low-level bit-fiddling and
string manipulation, but due to the configuration code that also
resides there, it depends on absolutely everything. This moves the
configuration code away, to keep the two worlds separated.

That said, shared.h itself already depends on everything, so it
doesn't really matter that much what shared.c itself contains - but I
might have a longer term plan for that, than merely this commit...

Change-Id: I5fcae3e47359f6f1c4db5f5c699017a056ac66df
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>