daemon: Remove live status data tracking
It's been deprecated for years, and it's in my way.

We still trigger ocimp to import timeperiods, because livestatus can't
export this data, and we still have to create *those* tables - but
they're per-naemon-start variables, so any and all live-updating is
gone.

For compatibility-like reasons, CTRL_PATHS still sends status.log from
the module, and the daemon simply ignores it.

Change-Id: Id0636e27bfa68d4d94953469443c83f4f06a979d
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>