codec: Fix handling of too long check output
When the check output is above 128kB (-ish), merlin tried to truncate
the output so it would fit into its protocol. However, it failed with
an off-by-one error, causing the next package(?) to be corrupt, and a
faux-reconnect to take place. This fixes the off-by-one, resolving
bug #9484.

Additionally, if the string that was too long wasn't the last string,
the subsequent strings weren't set to NULL, leading to ugly noise in
the daemon log:

 [2014-11-25 12:00:00 +0100] 3: Nulling OOB ptr 3. cb: NOTIFICATION;
 type: 601; offset: 0x13bad20; len: 131072; overshot with 20557088
 bytes

Since this patch also sets the subsequent strings to NULL, this should
also be taken care of. As the strings for check results are host name,
possibly service description, output, perf data, and long output, they
seem to be ordered roughly in order of importance - I'd rather loose
long output completely and get the whole host name, than getting half
of each.

Change-Id: I8838a55944d7c7130f6ad13f2e16f54ecd169de0
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>