oconf: Include read flag for other users in split config
Split config is always split as root:root, and then those permissions
are persisted across pushes to pollers. This means we need "others" to
have read access.

This is what the old, non-mkstemp code did, but I managed to goof up
the conversion, which means that, without this patch, you need
--use-helper to get a working system.

Change-Id: I8ba0897e89ad15a4df78f3c26e22eb0e420b5a2f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>