check.py: Don't crash when queryhandler returns funky data
It's possible that a peer id mismatch is due to missing/totally bogus
information about the other node. This shouldn't really happen, but if
it does, it would be good if we print *why* we report something,
rather than failing to print the error, stacktracing, and then, if the
check is run through NRPE, ignoring stderr (where the stacktrace is),
and just have the check alert with an "NRPE: Unable to read output" -
which totally doesn't help you pin down why you were texted at 3AM.

I think slash hope that this only happens with the broken 7.0 merlin,
but let's still avoid crashing in the check when it's trying to tell
you something important.

Change-Id: I50be79b3ca73b87f95be04e8a888388af5cb5fb0
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>