import.c: Stop faffing around with commands for downtimes
We never, ever use mon log import --truncate anymore. We use
incremental imports occasionally, though. For those, we need to
respect the damn log and insert what it contains - a start is a start,
a stop is a stop, commands be damned. We used to instead have a
complicated parsing of all types of nagios downtime commands, just so
we could match the downtime start event against that command, and if
we didn't find a matching command (because, I don't know, it was submitted
2 years ago, maybe?), we'd assume the downtime was 2 hours long. A
downtime end that didn't match a start was ignored. And yes, that
means that a downtime end that did match a start, but was more than 2
hours after the start and there was no command in scope means that we
consider the downtime end "unmatched".

This. Was wrong.

"What about tests?" you might say. There are a few unit-test like
tests for this binary in ninja's report tests (of all places) - after
manually installing this updated binary, they still pass just fine.

Change-Id: I6188cd64cccd4bc85c7e9324109c68282f759316
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>