mon sshkey push: Add known_hosts key for monitor user as well
... but only when pushing as root, obviously.

This means that we'll only ever ask to accept the remote hostkey
once during the exchange, and a single "mon sshkey push --all"
as root should be enough (from each master-peer, that is) to
ensure that non-interactive logins work from monitor to monitor,
from root to root and from root to monitor.

Are we done with #4481 now?

Change-Id: I6f5ee442195eea4a475df66311455b6287c68e01
Signed-off-by: Andreas Ericsson <ae@op5.se>