build: Create config cache directory
Without this, configsplit always fails, leading to sadness and broken
systems all around. The only thing that didn't fail without this was
the unit tests - as long as the oconfsplit completely failed, it
didn't ruin the internal naemon object structures, so the unittests
(which have a pre-split config) passed. Which they shouldn't, because
there are critical errors in the oconfsplit, as well.

Change-Id: I287fcb8f5cff5be33c7bedbea408541b413cb57e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>