Remove stalling
Stalling was kind of necessary when we used to track object status
in the database, where we didn't want to send any events to the
daemon before the import was fully completed. Since status tracking
in the database is no longer supported and the only thing we use
the import for is to get timeperiods into the database for our
reports, we no longer need to stall while the import is running.

This should fix one side of issue #9399.

Change-Id: I840b07f42d24da5730ce946def51dbfba356292c
Signed-off-by: Andreas Ericsson <ae@op5.se>