oconfsplit: Second half of "Stop manipulating global structures"
The first half was very, very broken and needed extra
verification. However, gerrit decided to merge it prematurely. Sigh.

This seemingly makes the initial approach work - that approach is to
try to avoid modifying the "live" config, and instead create a
temporary, non-public config - immutability is cool stuff. However, as
this commit shows, there is still interdependencies that makes this
approach dicey as well - creating a temporary host with a hostgroup
membership affects the hostgroup - so in the end, I have to reset all
the interdependent memory structures manually.

Change-Id: Ie97e7646615d16c0118d4e9952b54612e1966fbc
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>