Remove stalling
This will potentially cause issues for people who are tracking
object status in their databases. However, since that's no
longer supported (in the future, which is now, but on a separate
development track), we have no big qualms about that.

This is part of an effort to backport the fix for issue #9399 to
the 6.x track of monitor.

Change-Id: I23b3235d82a2b5ae470c605f87f11d987666a421
Signed-off-by: Andreas Ericsson <ae@op5.se>