module: Make sure reconnecting nodes count as reconnected
Normally, the CTRL_ACTIVE packet is the first we receive from a
disconnected node, so for the normal case this won't really be
an issue.

However, when a node disconnects for other reasons than going
down, and we then receive some backlogged data for it that
isn't a CTRL_ACTIVE packet, we would previously end up with a
node that's STATE_CONNECTED but that hasn't triggered a peer
id reassignment.

I'm a little hazy on why exactly that happened, but by
enforcing a connectivity chain where nodes are only ever
considered "connected" in one place, we fix that issue and
merlin clusters will now self-heal over some period of time.
That "some period" depends on what caused the issue, but
should be no longer than 45-60 seconds at worse.

As a part of this fix, we also make sure that we do the
peer-id reassignment thing when a node connects anew, even
if the data sent in the CTRL_ACTIVE packet is the same as
it sent the last time.

This fixes issue #9613.

Change-Id: I83bb97e9c321036ebece21b126a612e671a86f73
Signed-off-by: Andreas Ericsson <ae@op5.se>