Handle node state transition errors
The loggin part of this was exceptionally helpful in tracking
down issue #9613, which seems to crop up from time to time in
different shapes and sizes. Since it will only should trigger
a noticeable change  there's an actual problem, we sort of
want this.

As per alofgren's suggestions, we disconnect nodes that do
illegal transitions, as such transitions indicate real
problems. As per usual, &ipc is excepted from this rule.

Change-Id: I0f263620e9f8a954d565102d2ac1003cc17ebe21
Signed-off-by: Andreas Ericsson <ae@op5.se>