module: Don't wait for nodes to connect
Doing so means we have to use iobroker_poll(), which can
trigger a lot of events to come in through connected nodes,
which can cause us to try to send things, which means we
can end up in an infinite loop of iobroker_poll()'ing that
we barely get out of.

In addition, it adds a (potentially) long delay before we
start monitoring things, which we really shouldn't have there.

This patch should fix issue #9399.

Change-Id: I036a2b793b5db6f85eba12f3e4489dee69aa1983
Signed-off-by: Andreas Ericsson <ae@op5.se>