queries: Log incoming requests on DEBUG, not INFO
Since we make a lot of nodeinfo queries to check the number of
configured peers, it becomes a bit spammy when each of these queries are
logged on a level enabled by default. In addition, the information is
not too interesting to a regular user. Let the messages go to DEBUG
instead, since I can imagine they're only useful if you're debugging
requests specifically.

This fixes #9757.

Change-Id: I1e48456d8a29cb106790451b56b69dde34593ad6
Signed-off-by: Anton Lofgren <alofgren@op5.com>