Use correct size requirement when checking for available data
When we need to actually grab data from the bufferqueue after
having peek'ed the packet header, we must check if the queue
still has enough data for both the header (which we peeked) and
the packet body (which we didn't). Failing to do so leads us to
believe that we have a bug in our own code, which we handle by
disconnecting the remote node.

This should fix issue #MON-7970, which I *think* can only be
reproduced when merlin sends packets that are larger than
MTU - HDR_SIZE, or when it sends packets that have a body
which is smaller than HDR_SIZE (which I don't think we ever
do, but I might be wrong). The MTU - HDR_SIZE condition is
something we do not test and I have no idea how to make a
test that reliably fails when it should, since the network
card may just be in a good mood and send both packets
before we attempt to read them. Hence, no tests added for
this issue.

Change-Id: Ia2dc00b5f1eb67b3a081416b28b17f8b801120ed
Signed-off-by: Andreas Ericsson <ae@op5.se>