mon oconf push: Make sections to push individually selectable
This patch makes it possible to select only certain items to push
to remote nodes. The different sections are "oconf" (for object
configuration), "bsm" (for business services manager rules) and
"extras" (for everything mentioned in the "sync" block of the
configuration).

As a way of testing it, we add the "--dryrun" flag, which just
prints the rsync commands we would have run to synchronize
files and exits without actually pushing anything. We don't
do the same printout for the monitor reload command however,
but it's not executed. Automated tests are also added which
verify that a generated configuration with paths configured
for pushing generate valid rsync commands to be executed.
These tests are automagically run at every build. In order for
that to work properly, we also have to fix a particular
idiosynchrasy of the build system, where codefiles had some
of their variables substituted during './configure', while
others were substituted during 'make', and for some reason
in different ways.

This is part of the workaround we want to suggest for MON-7968.

Change-Id: I9b0d175b469341fd9c992e2d659768a58de5f276
Signed-off-by: Andreas Ericsson <ae@op5.se>