daemon: Be less aggressive when disconnecting peers
It's possible for peers to have configuration that is in
sync with ours but has a different timestamps. Without
this patch, peers that ended up in that state would never
be connected and would never push its configuration. Since
we really don't want to push when we know the config is
already identical, we go with the "allow the connection"
in those cases rather than the "disconnect and do nothing"
approach.

While at it, we get rid of the awkward magic numbers used
by handle_ctrl_active() and replace them with macros that
have at least some mnemonic meaning, and since we never
evaluated any combination of the negative numbers anyway,
we just convert them to somewhat "normal" error codes, of
which we test for and choose the most severe to return.

Part of the fix for MON-7968.

Change-Id: Ib50c64d2826a8ac43fa81ed1de5007b9947efce3
Signed-off-by: Andreas Ericsson <ae@op5.se>