daemon: Don't limit configpushing only to peers
If we get ESYNC_ECONFTIME back from the inspection of a
CTRL_ACTIVE packet, that means the node's object configuration
is older than ours. If that happens with any type of node, we
need to see if we should push our configuration to that node.

This patch makes sure we do for pollers, when it was previously
only done for peers. It has been manually verified to be working
as intended by myself and Philip Eklöf.

This fixes MON-8048.

Change-Id: Id455d63ad2cfd7f4c6fd9595d4c3d53a6ec03615
Signed-off-by: Andreas Ericsson <ae@op5.se>