oconfsplit: Don't overwrite active service
8c417d5778 was meant to cause oconfsplit to avoid manipulating global
structures. Unfortunately, it failed to use the "safe copy" of services
when splitting hosts and instead used the very instance it was meant to
avoid. This lead to the services being "reset" to their initial states on
every restart.

This patch fixes that, in addition to marking the structures we don't
want to mutate const. Now, merlin for whatever reason doesn't build with
-Werror enabled anywhere, but at least this particular slip-up will be
warned about.

This fixes MON-8123 and MON-8109.

Change-Id: I0c48ece6f8f5e5e0a851d7ed5e4bc3f3c0c3a42e
Signed-off-by: Anton Lofgren <alofgren@op5.com>