daemon: Make things build against dbi version 0.9 without warnings
We employ macros to turn the reentrant API into an identical
version of the old, non-reentrant one, which saves us a lot
of #ifdef/#else code inside the actual logic, since we also
want the code to build at all against dbi version 0.8.

This is the same solution previously applied for trapper,
except this one doesn't create the macro if it doesn't exist,
since other projects we might want to link against could
check to see if it's set in order to determine if libdbi is
version 0.9 or higher.

While at it, we remove a bunch of debug macros which were
used but commented out inside the code, since the macros
themselves were shelved alongside this solution.

Change-Id: Id44dfd090a229e6814030d3a3ae854632d01c31b
Signed-off-by: Andreas Ericsson <ae@op5.se>