Make database layer available to module as well
While we really would prefer to get rid of it altogether, the
first step right now is to make sure the module runs all the
helper scripts, since the module is the one that writes the
split configuration files and therefore know where they are
and can figure out which configuration a particular poller
should have. In order to run helpers, the module needs to know
the database configuration so it can pass it on to those
helpers.

While at it, we resolve MON-7820, as it pertains to a database
configuration value and we're fiddling in the area anyway.

Change-Id: I96c3639425188cd1290571d46f5915df7bd54922
Signed-off-by: Andreas Ericsson <ae@op5.se>