module: Set the expected config hash for pollers
Since we're splitting the configuration in the module anyway, we
know where a particular poller configuration file ends up, and we
can easily hash it to see which configuration we expect the poller
to have. In doing so, we make it trivial to verify if the poller
has the correct configuration or not during runtime.

This is part of the fix for MON-7969.

Change-Id: I173e24601b7e997565cf265996ad093cff8d4884
Signed-off-by: Andreas Ericsson <ae@op5.se>