janitor-ish: Squelch some compiler warnings
When refactoring, it's supremely annoying to see warnings you don't
care about, but most of them are actually sensible. In case we fail
to asprintf() some string, we'll most likely crash afterwards. The
depressing part is that asprintf() will almost certainly only fail
if we run out of memory, or the compiler would catch format string
errors anyway, so if we manage to log anything is anyone's guess.

However, we should at least try, and this patch does that, while
also fixing some other, less significant warnings (such as const
correctness and signedness warnings).

This makes refactoring necessary to fix MON-7969 easier.

Change-Id: Id7068c18709d870d2e0f832c4231d14f3b3fb541
Signed-off-by: Andreas Ericsson <ae@op5.se>