Refactor compatibility checks
It makes little sense to have such poor granularity over such
importance checks, so we split it into three separate functions,
with "the old way" (handle_ctrl_active()) of calling it retained
for backwards compatibility.

This will be used later in order to check compatibility between
daemon and module, as well as loosen the compatibility requirements
between daemons and also make sure we take the proper action
depending on which kind of incompatibility we run into.

Part of fixing MON-7969 and a slew of other issues.

Change-Id: I736166c097ef0e7092dcb729ef07792bd0be2807
Signed-off-by: Andreas Ericsson <ae@op5.se>