Move object config synchronization from daemon to module
The module is exceptionally well prepared to run external programs
asynchronously, and it also has the information about which config
each node *should* have, so let the module take care of the sync.

As a nice bonus, we also get output gathering for free with this
move, which means we can actually log both stdout and stderr output
from the config sync programs we run without really having to do
anything special in order to get it.

This goes a long way towards getting the suggested 7-step fix
decided on for MON-7969.

Change-Id: I180c7996397d98c82168c97e29a92569e0a24588
Signed-off-by: Andreas Ericsson <ae@op5.se>