Finetune compiler warnings a bit more
In particular, we don't want to see the errors from the return
value of asprintf() being ignored, and there are several other
similar functions we really can't do anything about.

However, in order to portably support the -Wno-unused-result
flag to gcc, we must make it conditional, so we import the
AX_CHECK_COMPILER_FLAG macro and use that, which lets us
finetune the flags we want without breaking the build on
systems that don't support all of them.

This makes refactoring easy, which is necessary for MON-7969.

Change-Id: Iab27115fc9a0888048be884343908a669c93cec0
Signed-off-by: Andreas Ericsson <ae@op5.se>