Fix a bunch of compiler warnings
Primarily redundant declarations for the module which we now
get from the naemon headers along with a whole slew of
shadowed and unused variables.

Declarations after statements has been disabled, however,
because the db_wrap_dbi code is more than just a little
messy in that regard. It will be enabled and fixed in the
next commit.

Change-Id: I00f89fb5197e2406500e1540c041bbf1afcd26fb
Signed-off-by: Andreas Ericsson <ae@op5.se>