mon test dist: Add the --no-confgen option and update helptext
The --no-confgen option makes it easier to test various
scenarios with customer configurations without the need to
set up virtual machines to cover all the possible cluster
setups our customers have. It does require the directories
one want to test to exist, but it's pretty nifty as it
provides a way to first create a configuration with

  mon test dist --confgen-only --pgroups=0 --masters=2

(for a peered setup, for example), modify the generated
configuration as necessary and then launch and control the
necessary daemons from a single command and terminal (namely
  mon test dist --no-confgen --pgroups=0 --masters=2 \
    --test=connections

As a part of this, we make the test_until_or_fail() function
private the way Python does such things, by prefixing it
with an underscore, since it otherwise shows up in the list
of available tests. We also shorten the original description
text a bit; We don't need to go into specifics regarding how
the tests test what they test.

Not really related to an issue, but it helps me test and set
up reproduction recipes for pretty much all merlin-related
ones (such as MON-7291) without having to keep several
virtual machines on my harddrive.

Change-Id: I45e636dc45d372c4a4dc03afc181b5686b75db34
Signed-off-by: Andreas Ericsson <ae@op5.se>