nodeinfo dump: Add info about config sync pushing and attempts
This makes it possible to see via query handler which nodes
we've attempted to sync config to, and how we tried to do it.
It also makes it trivial to verify that the config sync config
in itself is what we want it to be.

This makes it easier to test MON-7969.

Change-Id: I44e822764e30c18bab9b323e14a9d178e38e8fb1
Signed-off-by: Andreas Ericsson <ae@op5.se>