mon test dist: Test per-node config pushing explicitly
The former test verified that we had tried to push at all. That lead
to false positives, which in turn lead to object configuration not
being pushed to all the nodes it was supposed to be pushed to.

Since we now know via the query handler which nodes triggered a push
attempt we should make sure that all of them have. This test does
just that, making it possible to refactor a bit more wildly in order
to solve the sideffects (partial config pushing) introduced when
fixing issue MON-7969.

Change-Id: Iebe5cf70a3f8463b1a97cdc940aa0d0b9af23fc0
Signed-off-by: Andreas Ericsson <ae@op5.se>