janitor: Clarify a comment
"test magic" is usually bad to have in production code, but when
it's a standard trick and clearly documented it's less iffy.

We mention how we use the "if oconf sync command == : then do nothing",
where it's used from and a sort of half-excuse to make sure future
programmers know what's going on.

Change-Id: I96b83ae476d2a274de60cda52d02e79b4231d4ab
Signed-off-by: Andreas Ericsson <ae@op5.se>