Add code for converting nebevents to kvvecs and vice versa
This code needs some polish (such as making it read the structs from
a header file instead of having them hardcoded in the file itself),
but as it's going to be used primarily for testing we might as well
get it out there quickly.

For some reason I'm having issues getting GNU make to understand
the dependency chains, so the generated files are presently in
the magic .PHONY target which means they're always executed.

Buildrules are added, but the sources aren't linked to anything yet
as the code that will use this isn't written yet.

This fixes the "generation of neb event parsing" part of MON-8621,
as well as its inverse, MON-8622.

Change-Id: I18f0d3f53efe554d4d24daf79d1e84cb4e4f8ed5
Signed-off-by: Andreas Ericsson <ae@op5.se>

Conflicts:
	.gitignore