merlincat: Add linebuffered reading from console
To be able to parse events to send to merlin, we need to be able to
asynchronously read from the console, and read line buffered.

This creates a GUnixInputStream, which is polled as a GSource, and adds
a line buffer, so lines upp to 1M can be handled at once.

Part of MON-8621

Change-Id: I1bb5dafda2730f1c94a1cc180bacb00a97e6cdc7
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>