merlincat: Make it possible to send packets
Make packer able to pack packets, so we can send them on the wire.

This also changes the state of a merlin_event so that all merlin_events
is on wire-format, not unpacked. Thus, to access data in the merlin
event (like when /un/packing to kvvec), we need to pass it thorugh the
coder/decoder

Some cleanup is needed of output and merlincat.c though

Almost fixes MON-8621

Change-Id: Ie9b04961ff0c3e3396e6d2d0a72646c0b8ba33de
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>