merlincat: Be glib2-2.22-compatible
glib2-2.22 doesn't have support for GPollableInputStream, or
g_source_add_unix_fd. So we need to use g_source_add_poll, which has
some side effects. So create a GSource, with all boilerplate code for
GSource storage. Add a filter in the GSource, and don't guse the stream
argument in the GPollableInputStream callback.

In that case, the code can work for GPollableInputStream in the future
without to much problem

Change-Id: I8097f4059d582b395103ef9feed363ac7a6b8d54
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>