cukemerlin: Create daemon that answers to cucumber
To test merlin protocol, a daemon that answers to cucumber connections
is necessary. So implement a merlin cucumber test daemon

It can be used by adding a step definition file, named anything.wire,
containing an address and a port, and then run a feature with that step
definition.

It will however just answer "success" to everything.

Resolves: MON-8690

Change-Id: I16c4d6898b505281b1859557ad4a3dc04cf00a79
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>