cukemerlin: Cleanup API with defines
Step definitions has a predefined syntax, but we might want to add some
parameters to the functions later, if we change the internal interface
of how to return strings and extra data later.

So centralize the definition of steps, so we can update that later.

Also, to minimize risk of developer error, use a NULL-terminated list of
steps instead of a length-prefixed.

Refactor before MON-8615

Change-Id: I195eb943663342d761504e95903527e4fe3a0c8b
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>