cukemerlin: Add steps for matching received events
Half of a test is actually to measure and verify, so add steps to verify
the received data too, not only send packets.

"Then X received event EVENT_TYPE" followed by a table verifies that a
message of a given type has been received since last time the receive
buffer is cleared, and that the fields specified in the table matches.

It is vaild to get other events also during that time, and even
non-matching of the same type. Just that at least one message should
match is enough

Resolves MON-8692

Change-Id: Ibed03820f29c2ef14c37264f9c54f24a1f5de670
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>