Allow relative paths to IPC
It doesn't really make sense to have relative paths to the IPC in a
regular system, so normal users wont use it. But there are some special
cases it makes sense, and thus it doesn't really make sense to forbid it
either.

Those cases is for example:
- Test environments, where configurations are created in a temporary
  directory, and merlind/naemon is started within that directory
- Systems where configuration files is generated, or running multiple
  instances of the system using partly the same configuration, so parts
  of the configuration is dependent on working directory (start one
  merlind per work dir)

In this case, the first step the reason for this patch.

Part of MON-8813

Change-Id: I6a23422d2f203c910bc8141beaddf00333188608
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>